Noetic support for STDR Simulator installation
The script for installing the STDR Simulator has been added.  There is a fork hosted on the course GitHub repo that has an updated version and this script uses that repo for Noetic, but the STDR repo for Melodic.  The Noetic branch has been submitted as a pull request.

The script also attempts to use rosdep to install dependencies.  Melodic does not properly declare map_server currently, so it must be installed manually.  It is anticipated to generate a minor update for the Melodic branch that fixed the dependency issues.